+++ examples/polypharmacy_gripnet/utils.py
@@ -1,4 +1,3 @@
-import imp
 import os
 
 import numpy as np
@@ -75,44 +74,3 @@ def get_all_dataloader(data: Data):
         dataloader_list.append(loader)
 
     return dataloader_list
-
-
-# ----------------------------------------------------------------------------
-# Copy-paste from https://github.com/NYXFLOWER/GripNet
-
-
-def negative_sampling(pos_edge_index, num_nodes):
-    idx = pos_edge_index[0] * num_nodes + pos_edge_index[1]
-    idx = idx.to(torch.device("cpu"))
-
-    perm = torch.tensor(np.random.choice(num_nodes ** 2, idx.size(0)))
-    mask = torch.from_numpy(np.isin(perm, idx).astype(np.uint8))
-    rest = mask.nonzero().view(-1)
-    while rest.numel() > 0:  # pragma: no cover
-        tmp = torch.tensor(np.random.choice(num_nodes ** 2, rest.size(0)))
-        mask = torch.from_numpy(np.isin(tmp, idx).astype(np.uint8))
-        perm[rest] = tmp
-        rest = mask.nonzero().view(-1)
-
-    row, col = perm / num_nodes, perm % num_nodes
-    return torch.stack([row, col], dim=0).long().to(pos_edge_index.device)
-
-
-def typed_negative_sampling(pos_edge_index, num_nodes, range_list):
-    tmp = []
-    for start, end in range_list:
-        tmp.append(negative_sampling(pos_edge_index[:, start:end], num_nodes))
-    return torch.cat(tmp, dim=1)
-
-
-def auprc_auroc_ap(target_tensor, score_tensor):
-    y = target_tensor.detach().cpu().numpy()
-    pred = score_tensor.detach().cpu().numpy()
-    auroc, ap = metrics.roc_auc_score(y, pred), metrics.average_precision_score(y, pred)
-    y, xx, _ = metrics.precision_recall_curve(y, pred)
-    auprc = metrics.auc(xx, y)
-
-    return auprc, auroc, ap
-
-
-# ----------------------------------------------------------------------------
