+++ kale/evaluate/metrics.py
@@ -1,3 +1,7 @@
+import torch
+from sklearn import metrics
+
+
 def concord_index(y, y_pred):
     """
     Calculate the Concordance Index (CI), which is a metric to measure the proportion of `concordant pairs
@@ -21,3 +25,20 @@ def concord_index(y, y_pred):
         return total_loss / pair
     else:
         return 0
+
+
+def auprc_auroc_ap(target: torch.Tensor, score: torch.Tensor):
+    """
+    AUPRC: the area under the precision-recall curve
+    AUROC: the Area Under the Receiver Operating Characteristic curve
+    ap: average precision
+
+    Copy-paste from https://github.com/NYXFLOWER/GripNet
+    """
+    y = target.detach().cpu().numpy()
+    pred = score.detach().cpu().numpy()
+    auroc, ap = metrics.roc_auc_score(y, pred), metrics.average_precision_score(y, pred)
+    y, x, _ = metrics.precision_recall_curve(y, pred)
+    auprc = metrics.auc(x, y)
+
+    return auprc, auroc, ap
