+++ kale/prepdata/graph_negative_sampling.py
@@ -1,0 +1,30 @@
+import numpy as np
+import torch
+
+
+def negative_sampling(pos_edge_index, num_nodes):
+    # Copy-paste from https://github.com/NYXFLOWER/GripNet
+
+    idx = pos_edge_index[0] * num_nodes + pos_edge_index[1]
+    idx = idx.to(torch.device("cpu"))
+
+    perm = torch.tensor(np.random.choice(num_nodes ** 2, idx.size(0)))
+    mask = torch.from_numpy(np.isin(perm, idx).astype(np.uint8))
+    rest = mask.nonzero().view(-1)
+    while rest.numel() > 0:  # pragma: no cover
+        tmp = torch.tensor(np.random.choice(num_nodes ** 2, rest.size(0)))
+        mask = torch.from_numpy(np.isin(tmp, idx).astype(np.uint8))
+        perm[rest] = tmp
+        rest = mask.nonzero().view(-1)
+
+    row, col = perm / num_nodes, perm % num_nodes
+    return torch.stack([row, col], dim=0).long().to(pos_edge_index.device)
+
+
+def typed_negative_sampling(pos_edge_index, num_nodes, range_list):
+    # Copy-paste from https://github.com/NYXFLOWER/GripNet
+
+    tmp = []
+    for start, end in range_list:
+        tmp.append(negative_sampling(pos_edge_index[:, start:end], num_nodes))
+    return torch.cat(tmp, dim=1)
